feat: materials; fixed memory leak

+++ src/color.zig
@@ -15,9 +15,9 @@ pub fn write_color(comptime WriterType: type, out: WriterType, pixel_color: Colo
     g = @sqrt(scale * g);
     b = @sqrt(scale * b);
 
-    const ir: u32 = @floatToInt(u32, 256 * common.clamp(r, 0.0, 0.999));
-    const ig: u32 = @floatToInt(u32, 256 * common.clamp(g, 0.0, 0.999));
-    const ib: u32 = @floatToInt(u32, 256 * common.clamp(b, 0.0, 0.999));
+    const ir: u32 = @floatToInt(u8, 256 * common.clamp(r, 0.0, 0.999));
+    const ig: u32 = @floatToInt(u8, 256 * common.clamp(g, 0.0, 0.999));
+    const ib: u32 = @floatToInt(u8, 256 * common.clamp(b, 0.0, 0.999));
     
     try out.print("{} {} {}\n", .{ ir, ig, ib });
 }
